improve MPI usage with interoperability as goal
MPI interoperability is a useful goal of any library or runtime that
uses MPI.

There are two key changes here:

1) CAF tolerates out-of-band MPI initialization.  In the event that MPI
is initialized prior to entry into the CAF runtime init call, then CAF
will detect this and not take ownership of MPI (which merely means that
it will not call MPI_Finalize upon successful termination).

2) MPI_COMM_WORLD is never used by the CAF runtime except to duplicate
it.  This is absolutely critical for MPI interoperability since use of
MPI_COMM_WORLD can lead to false/cross matching of collectives and
send-recv matching.

Signed-off-by: Jeff Hammond <jeff.science@gmail.com>
Signed-off-by: Tobias Burnus <burnus@net-b.de>