Fix CMake problem with nonstandard version #
CMake apparently barfs on version a version number of the format "1.2.3-prerelease".  If we seek to distinguish git clones between releases from the released versions, then we will need to bump the version number both just after release and just before release. Starting from the current state, this will have the effect that version numbers ending in an even number are released versions and those ending in an odd number are between-release versions.