Handle missing wget gracefully.
The "build" script now detects a messing download mechanism (as designated in the $fetch variable) and prompts the user to download the relevant tar ball and restart the installation.  A few exit statements have also been updated to have unique exit codes. A minor edit in "install.sh" indicates to the user that the tar ball will be downloaded "if necessary", i.e., if it isn't already found in the expected location.