Non-interactive install options: -y/--yes-to-all
```
./install.sh -y
```
or equivalently
```
./install.sh --yes-to-all)
```
will now causes a non-interactive build by assuming affirmative
responses to all user queries.  Exception: if the user passes
an installation prefix (via -i or --install-prefix), the user
will be prompted for a password prior to the "make install" step.
To avoid this, execute as `sudo ./install.sh...` with the
consequence that all steps will execute with sudo privileges.