prevent silent failure of events integration test
src/tests/integration/events/async-hello.F90 was failing silently
with GCC 7.  This commit bumps the image count to 3 so the test
prints output that reports the fatal MPI error. This commit also
uses C preprocessor conditionals to prevent reporting success
for this test with GCC 7.  The pre-processor conditionals will
need to be removed manually when the silent failure has been
eliminated.