Fix sync images race condition.
Syncing a set of images using SYNC IMAGES is done by this simple algorithm:

1. Set up asynchronous mpi-receives for all images in the set to sync to,
2. Send the current image's status to all images in the set to sync to, and
3. Wait for the receives from step 1 to finish. When one of the receives
   returns a stopped image state, than abort the sync on the current image
   immediately, else wait until all receives have been answered.

This commit also adds a new testcase: syncimages_ring, where each image
is syncing to its neighbours with wrap around. The testcase is called for
3 (the minimum number of images required to show that the sync works), 13
and 23 images.

This commit fixes the timeout of syncimage_status.f90 mentioned in #298.
This commit superseeds pull request 299 (close #299).