Move MPICH path output later in the setup scripts
Because MPI is important of using OpenCoarrays, but CMake
is only important for building OpenCoarrays, we want to
ensure the MPI path precedes the CMake path in the PATH
variable in the setup scripts that the installer writes.
This lessens the likelihood of using a different MPI that
happens to be in the location of CMake (e.g., /usr/bin).