caf links opencoarrays module for caf compiler too
The opencoarrays module in src/extensions extends the capabilites
of a compiler (e.g., providing Fortran 2015 features for use with a
Fortran 2008 compiler).  The caf compiler wrapper now links the
opencoarrays module with caf compilers too (previously it was only
linked with non-caf compilers). This should only affect code that
explicitly imports module entities via "use opencoarrays".  In
case the compiler does support the features provided by the
opencoarrays module, it is important to append an ", only" clause
to the use statement to ensure that imported entities (e.g., co_sum)
don't clobber the compiler's implementation of the same feature
unless so desired.