CMake: fix -Wint-conversion warning (error if clang is used as MPIC_C)
Using of single quotes for printf argument within MPI_C check
of CMakeLists.txt results in warning of conversion to int.
The single quotes is used for char/int but not for char* type
where the double quotes should be used.

If clang is used as MPI_C instead of gcc the configuration is failed
with error:
incompatible integer to pointer conversion passing 'int'
to parameter of type 'const char *' [-Wint-conversion]

(refer Gentoo [guru overlay] issue: https://bugs.gentoo.org/888009)

Signed-off-by: Sergey Torokhov <torokhov-s-a@yandex.ru>