Prevent use of OpenMPI.
OpenMPI is incompatible with the way opencoarray uses MPI_wins. A
MPI_Win on OpenMPI is a pointer exclusive to each image. Because tokens
(aka MPI_Wins) are transfered between images, OpenMPI can not be used,
because no way of mapping exists. MPIch and Intel's MPI work fine.

This is not a bug, but a design decision.