mpi: exchange an index instead of MPI_Win
A requirement is that all the ranks of a given MPI window share the same id
(so it can be passed to an other rank). There is no such thing mandated by
the MPI standard, this is absolutely not the case with Open MPI (and its derivatives)
and though it seems to work with MPICH (and its derivatives) this is not something
guaranted by the implementation.

A reliable mechanism must be implemented for this to work with any MPI library.

This patch is a proof of concept that introduces two subroutines:
 - `MPI_Win CAF_idx2win(int idx)`
 - `int CAF_win2idx(MPI_Win win)`
that do the conversion between a MPI_Win opaque handler and and index that meet
the Coarrays requirements.

These subroutines are implemented with `MPI_Win_f2c()` and `MPI_Win_c2f()`.

THIS IS NOT A CORRECT IMPLEMENTATIONi
since the MPI Standard offer no guarantee
that the Fortran ID meet the requirements for Coarrays.

OpenCoarrays have been lucky so far with MPICH, and it may be equally lucky
with Open MPI too, but that CANNOT be taken for granted.

Refs #793

Signed-off-by: Gilles Gouaillardet <gilles@rist.or.jp>